pyproject.toml
[tool.poetry]
name = "eatchef_server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4"
argon2-cffi = "==20.1.0"
boto3 = "==1.17.86"
celery = "==5.1.2"
channels = "==3.0.3"
channels-redis = "==3.2.0"
django = "==3.2.4"
django-admin-json-editor = "==0.2.3"
django-celery-email = "==3.0.0"
django-cors-headers = "==3.7.0"
django-filter = "==2.4.0"
django-grappelli = "==2.15.1"
django-inline-actions = "==2.4.0"
django-nested-admin = "==3.3.3"
django-request-logging = "==0.7.3"
django-storages = {extras = ["google"], version = "*"}
djangorestframework = "==3.12.4"
djangorestframework-simplejwt = "4.8.0"
drf-yasg = "==1.20.0"
gunicorn = "==20.1.0"
jsonschema = "==3.2.0"
mutagen = "==1.45.1"
pandas = "==1.3.1"
phonenumbers = "==8.12.24"
pillow = "==8.2.0"
psycopg2-binary = "==2.9.1"
python-dateutil = "==2.8.1"
pyyaml = "==5.4.1"
redis = "==3.5.3"
requests = "==2.25.1"
sklearn = "*"
social-auth-app-django = "==4.0.0"
django-extensions = "^3.1.3"
PyJWT = "2.1.0"

[tool.poetry.dev-dependencies]
autopep8 = "==1.5.7"
coverage = "==5.5"
pytest = "==6.2.4"
tblib = "==1.7.0"
vcrpy = "==4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
